# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CategoryPostReq:
    properties:
      cate_name:
        type: string
    required:
    - cate_name
    type: object
  models.CategoryUpdateReq:
    properties:
      cate_name:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8888
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /category:
    get:
      consumes:
      - application/json
      description: get category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            type: ""
      security:
      - Bearer: []
      summary: List Category
      tags:
      - category
    post:
      consumes:
      - application/json
      description: Create New Category
      parameters:
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CategoryPostReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            type: ""
      security:
      - Bearer: []
      summary: Create New Category
      tags:
      - category
  /category/{id}:
    get:
      consumes:
      - application/json
      description: GetCategoryById
      parameters:
      - description: category id
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            type: ""
      summary: GetCategoryById
      tags:
      - category
    put:
      consumes:
      - application/json
      description: Update Category
      parameters:
      - description: category id
        in: path
        name: id
        type: integer
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CategoryUpdateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            type: ""
        "422":
          description: Unprocessable Entity
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            type: ""
      security:
      - Bearer: []
      summary: Update Category
      tags:
      - category
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"